Ints to short

+++ filestore/src/main/kotlin/dev/openrune/cache/filestore/definition/Recolourable.kt
@@ -4,18 +4,18 @@ import dev.openrune.cache.filestore.buffer.Reader
 import dev.openrune.cache.filestore.buffer.Writer
 
 interface Recolourable {
-    var originalColours: MutableList<Short>?
-    var modifiedColours: MutableList<Short>?
-    var originalTextureColours: MutableList<Short>?
-    var modifiedTextureColours: MutableList<Short>?
+    var originalColours: MutableList<Int>?
+    var modifiedColours: MutableList<Int>?
+    var originalTextureColours: MutableList<Int>?
+    var modifiedTextureColours: MutableList<Int>?
 
     fun readColours(buffer: Reader) {
         val length = buffer.readUnsignedByte()
         originalColours = MutableList(length) { -1 }
         modifiedColours = MutableList(length) { -1 }
         for (count in 0 until length) {
-            originalColours!![count] = buffer.readShort().toShort()
-            modifiedColours!![count] = buffer.readShort().toShort()
+            originalColours!![count] = buffer.readShort().toShort().toInt()
+            modifiedColours!![count] = buffer.readShort().toShort().toInt()
         }
     }
 
@@ -24,8 +24,8 @@ interface Recolourable {
         originalTextureColours = MutableList(length) { -1 }
         modifiedTextureColours = MutableList(length) { -1 }
         for (count in 0 until length) {
-            originalTextureColours!![count] = buffer.readShort().toShort()
-            modifiedTextureColours!![count] = buffer.readShort().toShort()
+            originalTextureColours!![count] = buffer.readShort().toShort().toInt()
+            modifiedTextureColours!![count] = buffer.readShort().toShort().toInt()
         }
     }
 
@@ -34,13 +34,13 @@ interface Recolourable {
         writeArray(writer, 41, originalTextureColours, modifiedTextureColours)
     }
 
-    private fun writeArray(writer: Writer, opcode: Int, original: List<Short>?, modified: List<Short>?) {
+    private fun writeArray(writer: Writer, opcode: Int, original: List<Int>?, modified: List<Int>?) {
         if (original != null && modified != null) {
             writer.writeByte(opcode)
             writer.writeByte(original.size)
             for (i in original.indices) {
-                writer.writeShort(original[i].toInt())
-                writer.writeShort(modified[i].toInt())
+                writer.writeShort(original[i])
+                writer.writeShort(modified[i])
             }
         }
     }
